requirements.txt
pip~=24.0
pillow~=10.2.0
Django~=5.0.4
asgiref~=3.8.1
sqlparse~=0.5.0
psycopg2~=2.9.9
numpy~=1.26.3
requests~=2.31.0
typing_extensions~=4.11.0
setuptools~=65.5.0
packaging~=23.2
pyparsing~=3.1.1
django-filter~=24.2

django-allauth~=0.63.1
certifi~=2023.11.17
urllib3~=2.1.0
idna~=3.6
pytz~=2024.1
tzlocal~=5.2
six~=1.16.0
APScheduler~=3.10.4